Call SetProgramName before initialising.
If SetProgramName is not set before the interpreter is initialised, then
the run-time libraries will be found relative to PATH. This isn't
necessarily what we want (e.g. see issues with cron in Trac #27085).



It's not clear what the 'right' name is here. For now, anything in
'/usr/local' will handle the PATH issue for the specific SpamExperts
case where the right Python to use is there. **This may not be
appropriate for any other users.**



Use bin/python as the name, as it seems safest to use a file that
actually exists, rather than something like 'local_scan'.